Change from storing vertices in a vector to storing them in a list
This facilitates adding a delete command which would be very slow if
the vertex descriptors were stored in a vector. This is because deleting
a vertex invalidates all descriptors so all references need to be reshuffled.
Using a list does not have this disadvantage so it gives us much more
flexibility.