Add aggregation for alarm suppression field.
The alarm suppression field should be aggregated to the
whole hierarchy. When the data is sent to the hierarchy service
for graph population, only model elements are expected to contain
this field. When data is queried from the cache, all nodes will
now contain this field. A parent node has the field set to true
if at least one of its children has it set to true. If not, it
is set to false.

This commit includes the following changes:

* Logic for aggregating the alarm suppression field for nodes queries.
* Tests verifying the logic implemented.
* Changes in hierarchy_dump, added field for suppression to model elements.